Moved to a domain based TLS white-listing.

+++ roku/channel_surfer.py
@@ -24,13 +24,14 @@ class SurferAborted(Exception):
 
 class ChannelSurfer(object):
 
-    def __init__(self, roku_ip, channel_id, data_dir, pcap_prefix, crawl_folder):
+    def __init__(self, roku_ip, channel_id, data_dir, pcap_prefix, crawl_folder, screenshot_folder):
 
         self.pcap_filename = None
         self.rrc = RokuRemoteControl(roku_ip)
         self.channel_id = str(channel_id)
         self.data_dir = data_dir
         self.pcap_dir = str(data_dir) + str(pcap_prefix)
+        self.screenshot_folder = str(data_dir) + str(screenshot_folder)
         self.go_home()
         self.log('Initialized', channel_id)
         self.crawl_folder = crawl_folder
@@ -152,6 +153,7 @@ class ChannelSurfer(object):
             int(timestamp)
         )
 
+        #self.log('./start_pcap.sh ' + str(self.pcap_dir) + "/" + str(self.pcap_filename))
         subprocess.Popen(
             './start_pcap.sh ' + str(self.pcap_dir) + "/" + str(self.pcap_filename),
             stdout=subprocess.PIPE, stderr=subprocess.PIPE, shell=True
@@ -164,7 +166,8 @@ class ChannelSurfer(object):
         start_time = time.time()
 
         while time.time() - start_time <= timeout:
-            screenshot_filename = '{}-{}'.format(self.pcap_filename, int(time.time()))
+            screenshot_filename = self.screenshot_folder + '{}-{}'.format(self.pcap_filename, int(time.time()))
+            #self.log('Taking screenshot to:', screenshot_filename)
             subprocess.call(
                 './capture_screenshot.sh {}'.format(screenshot_filename),
                 shell=True, stdout=subprocess.PIPE, stderr=subprocess.PIPE
@@ -172,12 +175,13 @@ class ChannelSurfer(object):
             self.deduplicate_screenshots(screenshot_filename)
 
     def deduplicate_screenshots(self, screenshot_filename):
-        screenshot_crc = binascii.crc32(open(screenshot_filename, 'rb').read())
-        if screenshot_crc == self.last_screenshot_crc:
-            self.log('Will remove duplicate screenshot:', screenshot_filename)
-            os.remove(screenshot_filename)
+        if os.path.exists(screenshot_filename):
+            screenshot_crc = binascii.crc32(open(screenshot_filename, 'rb').read())
+            if screenshot_crc == self.last_screenshot_crc:
+                self.log('Will remove duplicate screenshot:', screenshot_filename)
+                os.remove(screenshot_filename)
 
-        self.last_screenshot_crc = screenshot_crc
+            self.last_screenshot_crc = screenshot_crc
 
     def kill_all_tcpdump(self):
